refactor: Move files 'Game.hpp/cpp' to parent folder
